+++ samples/OwinSingleLogoutSample/Controllers/HomeController.cs
@@ -1,4 +1,7 @@
+using System.Threading.Tasks;
+using System.Web;
 using System.Web.Mvc;
+using Microsoft.Owin.Security.Cookies;
 
 namespace OwinSingleLogoutSample.Controllers
 {
@@ -6,9 +9,12 @@ namespace OwinSingleLogoutSample.Controllers
     public class HomeController : Controller
     {
         [HttpGet]
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
-            return View();
+            if (!User.Identity.IsAuthenticated) return View();
+            var result = await Request.GetOwinContext().Authentication
+                .AuthenticateAsync(CookieAuthenticationDefaults.AuthenticationType);
+            return View(result.Properties);
         }
     }
 }