+++ src/GSS.Authentication.CAS.Core/ServiceTicket.cs
@@ -12,8 +12,8 @@ namespace GSS.Authentication.CAS
             Assertion? assertion,
             IEnumerable<Claim> claims,
             string authenticationType,
-            DateTimeOffset? validFrom = null,
-            DateTimeOffset? validUntil = null)
+            DateTimeOffset? issuedUtc = null,
+            DateTimeOffset? expiresUtc = null)
         {
             if (string.IsNullOrWhiteSpace(ticketId)) throw new ArgumentNullException(nameof(ticketId));
             if (string.IsNullOrWhiteSpace(authenticationType))
@@ -24,15 +24,17 @@ namespace GSS.Authentication.CAS
 #pragma warning restore CS0618
             Claims = claims;
             AuthenticationType = authenticationType;
-            ValidFrom = validFrom;
-            ValidUntil = validUntil;
+            IssuedUtc = issuedUtc;
+            ExpiresUtc = expiresUtc;
         }
-        
+
         public ServiceTicket(string ticketId,
             IEnumerable<Claim> claims,
             string authenticationType,
-            DateTimeOffset? validFrom = null,
-            DateTimeOffset? validUntil = null)
+            DateTimeOffset? issuedUtc = null,
+            DateTimeOffset? expiresUtc = null,
+            string? nameClaimType = null,
+            string? roleClaimType = null)
         {
             if (string.IsNullOrWhiteSpace(ticketId)) throw new ArgumentNullException(nameof(ticketId));
             if (string.IsNullOrWhiteSpace(authenticationType))
@@ -40,21 +42,33 @@ namespace GSS.Authentication.CAS
             TicketId = ticketId;
             Claims = claims;
             AuthenticationType = authenticationType;
-            ValidFrom = validFrom;
-            ValidUntil = validUntil;
+            IssuedUtc = issuedUtc;
+            ExpiresUtc = expiresUtc;
+            NameClaimType = nameClaimType;
+            RoleClaimType = roleClaimType;
         }
 
         public string TicketId { get; }
 
         public string AuthenticationType { get; }
 
+        public IEnumerable<Claim> Claims { get; }
+
+        public string? NameClaimType { get; }
+
+        public string? RoleClaimType { get; }
+
+        public DateTimeOffset? IssuedUtc { get; }
+
+        public DateTimeOffset? ExpiresUtc { get; }
+
         [Obsolete("Use Claims instead. Will be removed in future release.")]
         public Assertion? Assertion { get; }
 
-        public IEnumerable<Claim> Claims { get; }
+        [Obsolete("Use IssuedUtc instead. Will be removed in future release.")]
+        public DateTimeOffset? ValidFrom => IssuedUtc;
 
-        public DateTimeOffset? ValidFrom { get; }
-
-        public DateTimeOffset? ValidUntil { get; }  
+        [Obsolete("Use ExpiresUtc instead. Will be removed in future release.")]
+        public DateTimeOffset? ValidUntil => ExpiresUtc;
     }
 }