+++ src/GSS.Authentication.CAS.DistributedCache/Internal/ServiceTicketHolder.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
@@ -13,7 +14,11 @@ namespace GSS.Authentication.CAS.Internal
         {
             TicketId = ticket.TicketId;
             AuthenticationType = ticket.AuthenticationType;
-            Claims = ticket.Claims.Select(x=> new ClaimHolder(x));
+            Claims = ticket.Claims.Select(x => new ClaimHolder(x));
+            NameClaimType = ticket.NameClaimType;
+            RoleClaimType = ticket.RoleClaimType;
+            IssuedUtc = ticket.IssuedUtc;
+            ExpiresUtc = ticket.ExpiresUtc;
         }
 
         public string TicketId { get; set; }
@@ -22,7 +27,16 @@ namespace GSS.Authentication.CAS.Internal
 
         public IEnumerable<ClaimHolder> Claims { get; set; }
 
+        public string? NameClaimType { get; set; }
+
+        public string? RoleClaimType { get; set; }
+
+        public DateTimeOffset? IssuedUtc { get; set; }
+
+        public DateTimeOffset? ExpiresUtc { get; set; }
+
         public static explicit operator ServiceTicket(ServiceTicketHolder h) =>
-            new ServiceTicket(h.TicketId, h.Claims.Select(x => (Claim)x), h.AuthenticationType);
+            new ServiceTicket(h.TicketId, h.Claims.Select(x => (Claim)x), h.AuthenticationType, h.IssuedUtc,
+                h.ExpiresUtc, h.NameClaimType, h.RoleClaimType);
     }
 }